Advance phase-2 header type-boundary cleanup
- Add explicit casts at key header accessors to reduce repeated conversion
  warnings from  return types.
- Update phase-2 status in the warning burndown plan with validated counts and
  net warning reduction so far.

Created by Copilot with gpt-5.3-codex

+++ trains/braid.h
@@ -13,7 +13,7 @@ public:
 	void Tighten() {Word.Tighten();}
 	uint Strings;
 	uint Size() {return Strings;}
-	uint Length() {return Word.TopIndex();}
+	uint Length() {return static_cast<uint>(Word.TopIndex());}
 	uint Permute(uint i); //Where does string i end up?
 	long& operator[] (uint i) {return Word[i];}
 	void Set(uint n, intarray W); //Sets n-braid from word W
@@ -27,4 +27,3 @@ public:
 
 
 #endif
-
